+++ src/mode/package/stats.ts
